rename parts of convertToHTML to better match their functions
