+++ test/spec/.eslintrc
@@ -1,0 +1,5 @@
+{
+  "env": {
+    "jasmine": true
+  }
+}
